rename seamless to seamless integration and remove shadow from range input
