+++ packages/docs/src/components/landing/SeamlessIntegration/SeamlessIntegrationAnotherButton.vue
