+++ packages/docs/src/components/landing/SeamlessIntegration/SeamlessIntegrationAnotherCheckbox.vue
