+++ packages/docs/src/pages/index.vue
@@ -12,7 +12,7 @@
       <Header />
       <Preview />
     </div>
-    <Seamless />
+    <SeamlessIntegration />
     <Customize />
     <OpenSource />
     <Admin />
@@ -28,7 +28,7 @@ import Preview from '@/components/landing/Preview.vue'
 import Admin from '@/components/landing/Admin.vue'
 import Footer from '@/components/landing/Footer.vue'
 import OpenSource from '@/components/landing/OpenSource.vue'
-import Seamless from '@/components/landing/Seamless/Seamless.vue'
+import SeamlessIntegration from '@/components/landing/SeamlessIntegration/SeamlessIntegration.vue'
 import Customize from '@/components/landing/Customize.vue'
 
 @Options({
@@ -38,7 +38,7 @@ import Customize from '@/components/landing/Customize.vue'
     Footer,
     OpenSource,
     Admin,
-    Seamless,
+    SeamlessIntegration,
     Customize,
   },
 })
